requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.20.0
attrs==25.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
nextcord==3.1.1
propcache==0.4.1
python-dotenv==1.0.1
requests==2.32.4
setuptools==75.8.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
